updated component structure of additional panel of market
