+++ packages/ui/config/index.ts
@@ -14,6 +14,7 @@ type CONFIG = {
   productDomain: string | undefined;
   productUrl: string | undefined;
   supabaseAssetApyTableName: string;
+  supabaseAssetPriceTableName: string;
   supabasePluginRewardsTableName: string;
   supabasePluginTableName: string;
   supabasePublicKey: string;
@@ -39,11 +40,12 @@ const config: CONFIG = {
   productDomain: process.env.PRODUCT_DOMAIN,
   productUrl: process.env.PRODUCT_URL,
   supabaseAssetApyTableName: process.env.SUPABASE_ASSET_APY_TABLE_NAME ?? '',
+  supabaseAssetPriceTableName: process.env.SUPABASE_ASSET_PRICE_TABLE_NAME ?? '',
   supabasePluginRewardsTableName: process.env.SUPABASE_PLUGIN_REWARDS_TABLE_NAME ?? '',
   supabasePluginTableName: process.env.SUPABASE_PLUGIN_TABLE_NAME ?? '',
   supabasePublicKey: process.env.SUPABASE_KEY ?? '',
   supabaseUrl: process.env.SUPABASE_URL ?? '',
-  supabaseVaultApyTableName: process.env.SUPABASE_VAULT_APY_TABLE_NAME ?? ''
+  supabaseVaultApyTableName: process.env.SUPABASE_VAULT_APY_TABLE_NAME ?? '',
 };
 
 export { config };
